requirements.txt
fastapi
uvicorn
torch
transformers
pillow
opencv-python
pytesseract
numpy
requests
zipfile